ContentForge/requirements.txt
requests>=2.31.0
rich>=13.7.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
